pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Awais", email = "muhammadawais9189@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "pytube>=15.0.0",
    "yt-dlp>=2024.5.26",
    "uvicorn>=0.34.2",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
]
